requirements.txt
pytesseract
pdf2image
pymupdf
Pillow
flask
werkzeug
